Billing GUI shall allow to reorganise the layout (#1100) - panels layout component & utilities refactoring
