+++ client/src/components/main/home/ConfigureHomePage.js
@@ -22,10 +22,7 @@ import {
   PanelIcons,
   PanelTitles,
   PanelInfos,
-  getPanelsLayout,
-  removePanel,
-  addPanels,
-  restoreDefaultLayout
+  Layout
 } from './layout';
 import {Button, Checkbox, Col, Icon, Modal, Row, Tooltip} from 'antd';
 import {getDisplayOnlyFavourites, setDisplayOnlyFavourites} from './utils/favourites';
@@ -48,17 +45,17 @@ export default class ConfigureHomePage extends localization.LocalizedReactCompon
 
   onSave = () => {
     const visibleKeys = this.state.panels.filter(panel => panel.visible).map(panel => panel.key);
-    const panels = getPanelsLayout();
+    const panels = Layout.getPanelsLayout();
     const removedPanels = panels.filter(item => visibleKeys.indexOf(item.i) === -1).map(item => item.i);
     const addedPanels = visibleKeys.filter(key => panels.filter(item => item.i === key).length === 0);
-    removedPanels.forEach(removePanel);
-    addPanels(addedPanels);
+    removedPanels.forEach(panel => Layout.removePanel(panel));
+    Layout.addPanels(addedPanels);
     setDisplayOnlyFavourites(this.state.displayOnlyFavourites);
     this.props.onSave && this.props.onSave();
   };
 
   restoreDefaultLayoutClicked = () => {
-    restoreDefaultLayout();
+    Layout.restoreDefaultLayout();
     this.props.onSave && this.props.onSave();
   };
 
@@ -150,7 +147,7 @@ export default class ConfigureHomePage extends localization.LocalizedReactCompon
 
   updatePanelsState = () => {
     const panels = [];
-    const layout = getPanelsLayout();
+    const layout = Layout.getPanelsLayout();
     for (let key in Panels) {
       if (Panels.hasOwnProperty(key)) {
         let title = PanelTitles[Panels[key]];
