+++ client/src/components/main/home/HomePage.js
@@ -19,11 +19,7 @@ import {inject, observer} from 'mobx-react';
 import GridLayout from 'react-grid-layout';
 import HomePagePanel from './HomePagePanel';
 import ConfigureHomePage from './ConfigureHomePage';
-import {
-  GridStyles,
-  getPanelsLayout,
-  setPanelsLayout
-} from './layout';
+import {GridStyles, Layout} from './layout';
 import {Button, Icon, Row} from 'antd';
 import PipelineRunFilter from '../../../models/pipelines/PipelineRunSingleFilter';
 import PipelineRunServices from '../../../models/pipelines/PipelineRunServices';
@@ -97,7 +93,7 @@ export default class HomePage extends React.Component {
   };
 
   onLayoutChanged = (layout, update = false) => {
-    setPanelsLayout(layout, false);
+    Layout.setPanelsLayout(layout, false);
     if (update) {
       this.forceUpdate();
     }
@@ -125,7 +121,7 @@ export default class HomePage extends React.Component {
     if (!this.props.authenticatedUserInfo.loaded && this.props.authenticatedUserInfo.pending) {
       return <LoadingView />;
     }
-    const panelsLayout = getPanelsLayout();
+    const panelsLayout = Layout.getPanelsLayout();
     return (
       <div
         ref={this.initializeContainer}
@@ -153,7 +149,7 @@ export default class HomePage extends React.Component {
             this.state.container &&
             <GridLayout
               className="layout"
-              draggableHandle={`.${GridStyles.draggableHandle}`}
+              draggableHandle={`.${styles.panelHeader}`}
               layout={panelsLayout}
               cols={GridStyles.gridCols}
               width={this.state.containerWidth - GridStyles.scrollBarSize}
