+++ client/src/components/main/home/layout/default-panels-sizes.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2019 EPAM Systems, Inc. (https://www.epam.com/)
+ * Copyright 2017-2020 EPAM Systems, Inc. (https://www.epam.com/)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-import styles from '../HomePage.css';
+import Panels from './panels';
 
-export const GridStyles = {
-  draggableHandle: styles.panelHeader,
-  gridCols: 24,
-  gridRows: 24,
-  padding: 0,
-  panelMargin: 5,
-  top: 40,
-  rowHeight: function (containerHeight) {
-    return (containerHeight - 2 * this.padding - this.gridRows * this.panelMargin - this.top) /
-      this.gridRows;
-  },
-  scrollBarSize: 15
+export default {
+  [Panels.activities]: {w: 2, h: 2},
+  [Panels.data]: {w: 2, h: 1},
+  [Panels.notifications]: {w: 1, h: 2},
+  [Panels.personalTools]: {w: 1, h: 2},
+  [Panels.pipelines]: {w: 1, h: 1},
+  [Panels.projects]: {w: 1, h: 2},
+  [Panels.recentlyCompletedRuns]: {w: 1, h: 2},
+  [Panels.runs]: {w: 1, h: 1},
+  [Panels.services]: {w: 1, h: 1}
 };
