+++ client/src/components/main/home/layout/default-panels-state.js
@@ -1,0 +1,21 @@
+/*
+ * Copyright 2017-2020 EPAM Systems, Inc. (https://www.epam.com/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export default [
+  {'w': 12, 'h': 12, 'x': 0, 'y': 12, 'i': 'my data', 'moved': false, 'static': false},
+  {'w': 12, 'h': 24, 'x': 12, 'y': 0, 'i': 'personal tools', 'moved': false, 'static': false},
+  {'w': 12, 'h': 12, 'x': 0, 'y': 0, 'i': 'my runs', 'moved': false, 'static': false}
+];
