+++ client/src/components/special/grid-layout/gridStyle.js
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2017-2019 EPAM Systems, Inc. (https://www.epam.com/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export function buildGridStyle (
+  {
+    cols = 24,
+    rows = 24,
+    padding = 0,
+    panelMargin = 5,
+    top = 40,
+    maxLayoutColumns = 4,
+    scrollBarSize = 15
+  }
+) {
+  return {
+    gridCols: cols,
+    gridRows: rows,
+    padding,
+    panelMargin,
+    top,
+    maxLayoutColumns,
+    rowHeight: function (containerHeight) {
+      return (containerHeight - 2 * this.padding - this.gridRows * this.panelMargin - this.top) /
+        this.gridRows;
+    },
+    scrollBarSize
+  };
+}
